requirements.txt
cryptography==38.0.4
Django==4.1.2
django-countries==7.5.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
lxml==4.9.1
MouseInfo==0.1.3
mysql-connector-repackaged==0.3.1
outcome==1.2.0
psutil==5.9.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyimgur==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-bcrypt==0.3.2
python-dotenv==0.21.1
requests==2.28.1
tao-utils @ git+https://github.com/taotradingltd/tao-utils.git@6f94a3e04be9e01b45c1bf689b31d776e5303dd0

